% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_sources.R
\name{check_data_sources}
\alias{check_data_sources}
\title{Check that data_sources are complete}
\usage{
check_data_sources(path = NULL)
}
\arguments{
\item{path}{option string giving the path for the root of the repository (NOT the data
directory). If left NULL, then the current working directory is used.}
}
\description{
Check that all directories in /data/raw and /data/intermediate are found in data_sources.csv,
and that their details are complete.
}
