% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dir_ignore.R
\name{data_dir_ignore}
\alias{data_dir_ignore}
\title{Untrack a data directory with git}
\usage{
data_dir_ignore(path)
}
\arguments{
\item{data_path}{the path of the data directory, given as relative to the root
of the git repository (e.g. "/data/raw/my_new_dir")}
}
\description{
This function removes tracking for a data directory in git.
}
