% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dir_link.R
\name{data_dir_link}
\alias{data_dir_link}
\title{Create a symlink to a data directory in a resrepo}
\usage{
data_dir_link(link_dir, target_dir)
}
\arguments{
\item{link_dir}{the link, given as relative to the root
of the git repository (e.g. "/data/raw/my_new_dir"). If the directory exists,
it must be empty, as it will be deleted and removed.}

\item{target_dir}{path outside the git repository that we want to link to
(i.e. where the data will be really stored)}
}
\description{
Create a symlink to a data directory in a resrepo (more details needed)
}
