% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dir_track.R
\name{data_dir_track}
\alias{data_dir_track}
\title{Track a data directory with git}
\usage{
data_dir_track(path)
}
\arguments{
\item{data_path}{the path of the data directory, given as relative to the root
of the git repository (e.g. "/data/raw/my_new_dir")}
}
\description{
This function sets up a data directory to be tracked
}
