% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_add.R
\name{data_source_add}
\alias{data_source_add}
\title{Add a new data source}
\usage{
data_source_add(dir = NULL, source = NA, url = NA, notes = NA, git_path = NULL)
}
\arguments{
\item{dir}{directory of the new entry. It should be given relative to
the root of the git repository. If the directory does not exist, it will be
created.}

\item{source}{source for the new entry (e.g. zenodo, onedrive, dropbox, etc.).}

\item{url}{url for the source.}

\item{notes}{notes for the new entry.}

\item{git_path}{option string giving the path for the root of the repository (NOT the data
directory). If left NULL, then the current working directory is used.}
}
\value{
TRUE if successful
}
\description{
This function adds a data source to the data_source_list.csv file. If the
directory does not exist yet, it will create one.
}
