% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_check.R
\name{data_source_check}
\alias{data_source_check}
\title{Check that data_sources are complete}
\usage{
data_source_check(path = NULL)
}
\arguments{
\item{path}{option string giving the path for the root of the repository (NOT the data
directory). If left NULL, then the current working directory is used.}
}
\description{
Check that all directories in /data/raw and /data/intermediate are found in data_source_list.csv,
and that their details are complete.
}
