% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_delete.R
\name{data_source_delete}
\alias{data_source_delete}
\title{Delete data source}
\usage{
data_source_delete(dir = NULL, git_path = NULL)
}
\arguments{
\item{dir}{directory of the new entry. It should be given relative to
the root of the git repository. If the directory does not exist, it will be
created.}

\item{git_path}{option string giving the path for the root of the repository (NOT the data
directory). If left NULL, then the current working directory is used.}
}
\value{
TRUE if successful
}
\description{
This function deletes a data source to the data_source_list.csv file. If still
present, the directory is removed, but only if it is empty.
}
