% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_list_add.R
\name{data_source_list_add}
\alias{data_source_list_add}
\title{Add one entry to existing data_source_list.csv}
\usage{
data_source_list_add(
  path = NULL,
  directory = NULL,
  source = "NA",
  url = "NA",
  notes_text = "NA"
)
}
\arguments{
\item{path}{option string giving the path for the root of the repository (NOT the data
directory). If left NULL, then the current working directory is used.}

\item{directory}{directory of the new entry.}

\item{source}{source for the new entry.}

\item{url}{url for the new entry.}

\item{notes}{notes for the new entry.}
}
\description{
Checks that data_source_list exists,
and if so, add user_specified entry into this csv file.
}
