% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_edit.R
\name{data_sources_edit}
\alias{data_sources_edit}
\title{Check that data_sources are complete}
\usage{
data_sources_edit(
  path = NULL,
  directory = NULL,
  new_directory = NULL,
  new_source = "NA",
  new_url = "NA",
  notes_text = "NA"
)
}
\arguments{
\item{path}{option string giving the path for the root of the repository (NOT the data
directory). If left NULL, then the current working directory is used.}
}
\description{
Check that all directories in /data/raw and /data/intermediate are found in data_sources.csv,
and that their details are complete.
}
