% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_git_root.R
\name{find_git_root}
\alias{find_git_root}
\title{Find the root of a git repository}
\usage{
find_git_root(path = ".")
}
\arguments{
\item{path}{option string giving the path to use a starting point (i.e. that
is part of the repository)}
}
\description{
Returns the path of the current git repository
}
