% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_resrepo_dir.R
\name{find_resrepo_dir}
\alias{find_resrepo_dir}
\title{Find the path to a certain directory in a resrepo}
\usage{
find_resrepo_dir(directory)
}
\arguments{
\item{directory}{directory within a resrepo}
}
\description{
Given a directory of a resrepo from its root (e.g. \verb{/data/raw/default}), it returns
the full path to it
}
