% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_version.R
\name{get_file_version}
\alias{get_file_version}
\title{Get a file from a specific version of the data}
\usage{
get_file_version(file_path, version = NULL)
}
\arguments{
\item{file_path}{The path to the file in the version of the data you want to
get. This path should be relative to the root of the repository (e.g.
"data/raw/original/my_file.cvs").}

\item{version}{The version of the data you want to get the file from. If
NULL, the current version is used.}
}
\value{
The path to the file in the current version of the data.
}
\description{
This function gets a file from a specific version of the data. It is useful
for comparing files between versions.
}
