% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resprepo_git_creds.R
\name{get_resrepo_git_creds}
\alias{get_resrepo_git_creds}
\title{Get git credentials if they are available}
\usage{
get_resrepo_git_creds()
}
\value{
A list containing the type of credentials, the URL, and the
GitHub PAT if applicable.
}
\description{
This function checks the current git remote URL and determines if the
credentials are set up for SSH or HTTPS access.
}
\keyword{internal}
