% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_is_clean.R
\name{git_is_clean}
\alias{git_is_clean}
\title{Check that a git repository has a clean working directory}
\usage{
git_is_clean(repo = ".")
}
\arguments{
\item{repo}{a path to a repository or a git_repository
object. Default is '.'}
}
\value{
TRUE if it has a clean working directory
}
\description{
Check that a git repository has a clean working directory
}
\keyword{internal}
