% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_resrepo.R
\name{init_resrepo}
\alias{init_resrepo}
\title{Initialise a resrepo}
\usage{
init_resrepo(path = ".")
}
\arguments{
\item{path}{optional path for the a resrepo; by defalt, we use the current
working directory}
}
\description{
Initialise a resrepo, creating all the necessary files and directories.
}
