% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data_dir.R
\name{is_data_dir}
\alias{is_data_dir}
\title{Check that a string is a valid datadir}
\usage{
is_data_dir(path)
}
\arguments{
\item{path}{string of path}
}
\description{
Internal function to check validtiy
}
\keyword{internal}
