% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_to_results.R
\name{knit_to_results}
\alias{knit_to_results}
\title{Custom knit function for RStudio}
\usage{
knit_to_results(inputFile, encoding)
}
\arguments{
\item{inputFile}{The name of the R Markdown file to be rendered.}

\item{encoding}{As per \code{rmarkdown::render_site()} this is ignored.
The encoding is always assumed to be UTF-8.}
}
\description{
This function is used in the YAML of an RMarkdown file so that when it is
knitted, the output \code{.md} file (and associated image files) are sent to
the \code{~/results} folder rather than the folder in which the \code{.Rmd}
file is located.
}
