% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_resrepo_dir.R
\name{make_datadir}
\alias{make_datadir}
\title{Make a data directory within a resrepo}
\usage{
make_datadir(directory)
}
\arguments{
\item{directory}{directory within a resrepo}
}
\description{
Make a directory within a resrepo with a short path from its root
(e.g. \verb{/data/raw/remote_sensing}). An empty directory will be added to the git
repository by addign a .gitkeep file to it, but .gitingore is set to avoid tracking
any other content. In this way, the resrepo will show all the primary level
directories in \verb{/data/raw} and \verb{/data/intermediate}, making it easier to keep
track of what data need downloading.
}
