% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_path_relative.R
\name{make_path_relative}
\alias{make_path_relative}
\title{Convert a path to relative}
\usage{
make_path_relative(path, base = getwd())
}
\arguments{
\item{path}{the full path}

\item{base}{optional, the base with respect to which the relative path is built. It defaults
to the current working directory in R.}
}
\description{
Given an absolute path, it returns the relative path with respect to a base
(by default, the working directory in R)
}
