% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_resrepo.R
\name{path_resrepo}
\alias{path_resrepo}
\title{Convert a path from its relative position within a resrepo}
\usage{
path_resrepo(path, check_exists = FALSE)
}
\arguments{
\item{path}{directory within a resrepo}

\item{check_exists}{boolean whether we should check if it exists}
}
\description{
Given a path relative to the root of a resrepo (e.g. \verb{/data/raw/default}), it returns
the full path to it. It works both on files and directories
}
