% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resrepo-package.R
\docType{package}
\name{resrepo-package}
\alias{resrepo}
\alias{resrepo-package}
\title{resrepo: Easy git repo management for reproducible science}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A package to manage git repositories used for research that contain a with a mix of code and data. resrepo is based on a well documented template with clear guidelines on how the repository should be structured, what file naming conventions to use, and how to manage large data that can not be stored in git repositories, encouraging reproducible science. In particular, resrepo provides tools to ensure that data are available and up-to-date both locally and in the cloud, as well as simplifying the process of locating directories within the repository from R scripts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://evolecolgroup.github.io/resrepo/}
}

}
\author{
\strong{Maintainer}: Andrea Manica \email{am351@cam.ac.uk}

Other contributors:
\itemize{
  \item Evie Carter [contributor]
  \item Andrea Vittorio Pozzi [contributor]
  \item Margherita Colucci [contributor]
  \item Cecilia Padilla-Iglesias [contributor]
  \item Michela Leonardi [contributor]
  \item Paul Whitelaw [contributor]
  \item Eirlys Tysall [contributor]
  \item Ondrej Pelanek [contributor]
  \item Zhe Xue [contributor]
  \item Anahit Hovhannisyan [contributor]
  \item Dominik Jud [contributor]
  \item Cassandra Gunasekaram [contributor]
  \item Elizabeth Pearmain [contributor]
  \item James Kennerley [contributor]
}

}
\keyword{internal}
