% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_links.R
\name{update_links}
\alias{update_links}
\title{Update links to the data directories}
\usage{
update_links(quiet = FALSE)
}
\arguments{
\item{quiet}{if TRUE, do not print messages}
}
\value{
TRUE if successful
}
\description{
Update links to the data directories, so that they point to the correct
version in the \code{version_resources} directory. If they links do not exist,
they will be created. If they exist, they will be updated only if the target
directory has changed.
}
