% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_add.R
\name{version_add}
\alias{version_add}
\title{Add a version of the data directories}
\usage{
version_add(
  path = ".",
  intermediate_new_version,
  raw_new_version = NULL,
  intermediate_source = NULL,
  raw_source = NULL,
  intermediate_description,
  raw_description = NULL,
  git_branch = NULL,
  quiet = FALSE
)
}
\arguments{
\item{path}{The path to the resrepo directory}

\item{git_branch}{The name of the git branch to create.
If NULL, it will be the same as \code{new_version}}

\item{quiet}{If TRUE, suppress messages}

\item{new_version}{The name of the new version}

\item{source_version}{The name of the version to copy from. If NULL, it will be the current version}

\item{description}{A description of the new version}
}
\value{
TRUE if the version was successfully added
}
\description{
This function will create a new version of the data directories, by copying the current
data directories to a new directory in \code{versions}, and creating links to the new
directories. The new version will be called \code{version_name}.
}
\details{
\code{intermediate_new_version} and \code{raw_new_version} will be sanitised to remove spaces and other characters that
are not allowed in directory names. If the sanitised version is different from
the original version, a warning will be issued.
}
