% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_exists.R
\name{version_exists}
\alias{version_exists}
\title{Check that a version exists in version_resources}
\usage{
version_exists(version, quiet = TRUE)
}
\arguments{
\item{version}{The version to check}

\item{quiet}{Should the function be quiet?}
}
\value{
TRUE if the version exists, FALSE otherwise
}
\description{
Check that a version exists in version_resources
}
\keyword{internal}
