% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_relink.R
\name{version_relink}
\alias{version_relink}
\title{Relink versions in a resrepo repository}
\usage{
version_relink(quiet = FALSE, resources_path = NULL)
}
\arguments{
\item{quiet}{If TRUE, the user will not be prompted to backup their data.
Default is FALSE.}

\item{resources_path}{The path to the directory where the "versions", the
directory containing versioned data, will be stored. If NULL (the default),
"versions" is placed at the root of the repository.}
}
\value{
TRUE if the relink was successful
}
\description{
If a repository has been cloned, relinks the "versions" directory to the
"data" directory according to the version metadata. If a user wishes to move
the "versions" directory to a different location, this function will also
update the links accordingly.
}
