% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_relink.R
\name{version_reset}
\alias{version_reset}
\title{Reset versions}
\usage{
version_reset(quiet = FALSE, resources_path = NULL)
}
\arguments{
\item{quiet}{If TRUE, the user will not be prompted to backup their data.
Default is FALSE.}

\item{resources_path}{The path to the directory where the "versions", the
directory containing versioned data, will be stored. If NULL (the default),
"versions" is placed at the root of the repository.}
}
\value{
TRUE if the relink was successful
}
\description{
Reset versions
}
\keyword{internal}
