% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_setup.R
\name{version_setup}
\alias{version_setup}
\title{Set up versioning for data in a resrepo}
\usage{
version_setup(quiet = FALSE, resources_path = NULL)
}
\arguments{
\item{quiet}{If TRUE, the user will not be prompted to backup their data.
Default is FALSE.}

\item{resources_path}{The path to the directory where the "versions",
the directory containing versioned data, will be stored. If NULL (the default),
"versions" is placed at the root of the repository.}
}
\value{
TRUE if the setup was successful
}
\description{
Move over the data to a separate directory, "versions", and create
links to the data in the repository. The first version of the data is, by
default, called "initial". Later on, we can add further versions of the
data, using \code{\link[=version_add]{version_add()}}.
}
