% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dir_follow.R, R/data_dir_track.R
\name{data_dir_follow}
\alias{data_dir_follow}
\title{Track a data directory with git}
\usage{
data_dir_follow(path)

data_dir_follow(path)
}
\arguments{
\item{dir}{the path of the data directory, given as relative to the root
of the git repository (e.g. "/data/raw/my_new_dir")}
}
\value{
TRUE if the directory is tracked

TRUE if the directory is tracked
}
\description{
This function sets up a data directory to be tracked by git. It does so by
adding the appropriate directory exception to \code{.gitignore}.

This function sets up a data directory to be tracked by git. It does so by
adding the appropriate directory exception to \code{.gitignore}.
}
