% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dir_make.R
\name{data_dir_make}
\alias{data_dir_make}
\title{Make a data directory within a resrepo}
\usage{
data_dir_make(path, source = NULL, URL = NULL, tracked = FALSE)
}
\arguments{
\item{path}{data path within a resrepo}

\item{source}{optional string giving the source of this data directory (e.g. onedrive)}

\item{URL}{optional string with the URL for the source}

\item{tracked}{boolean definign whether the directory should be tracked in git (this
will modify .gitignore appropriately)}
}
\description{
Make a directory within a resrepo with a short path from its root
(e.g. \verb{/data/raw/remote_sensing}). An empty directory will be added to the git
repository by addign a .gitkeep file to it, but .gitingore is set to avoid tracking
any other content. In this way, the resrepo will show all the primary level
directories in \verb{/data/raw} and \verb{/data/intermediate}, making it easier to keep
track of what data need downloading.
}
