% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_edit.R
\name{data_source_edit}
\alias{data_source_edit}
\title{Edit the data_source file}
\usage{
data_source_edit(
  directory,
  path = NULL,
  new_directory = NULL,
  new_source = NULL,
  new_url = NULL,
  new_notes = NULL
)
}
\arguments{
\item{directory}{the directory you want to change; has to be already in the data_source.csv}

\item{path}{option string giving the path for the root of the repository (NOT the data
directory). If left NULL, then the current working directory is used.}

\item{new_directory}{the new directory you want to give; if NULL then does not change.}

\item{new_source}{the new source you want to give; if NULL then does not change.}

\item{new_url}{the new url you want to give; if NULL then does not change.}

\item{new_notes}{the new dnote you want to give; if NULL then does not change.}
}
\description{
Edit the existing lines of data_source file.
}
