% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_add.R
\name{version_add}
\alias{version_add}
\title{Add a version of the data directories}
\usage{
version_add(
  path = ".",
  new_version,
  source_version = NULL,
  description,
  git_branch = NULL
)
}
\arguments{
\item{new_version}{The name of the new version}

\item{source_version}{The name of the version to copy from. If NULL, it will be the current version}

\item{description}{A description of the new version}

\item{git_branch}{The name of the git branch to create. If NULL, it will be the same as \code{new_version}}
}
\value{
TRUE if the version was successfully added
}
\description{
This function will create a new version of the data directories, by copying the current
data directories to a new directory in \code{version_resources}, and creating links to the new
directories. The new version will be called \code{version_name}.
}
