% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_setup.R
\name{version_setup}
\alias{version_setup}
\title{Set up versioning for data in a resrepo}
\usage{
version_setup(quiet = FALSE)
}
\arguments{
\item{quiet}{If TRUE, the user will not be prompted to backup their data.
Default is FALSE.}
}
\value{
TRUE if the setup was successful
}
\description{
Move over the data to a separate directory, "version_resources", and create
links to the data in the repository. The first version of the data is, by
default, called "initial". Later on, we can add further versions of the data,
and switch between them.
}
