% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_setup.R
\name{version_setup}
\alias{version_setup}
\title{Set up versioning for data in a resrepo}
\usage{
version_setup()
}
\description{
Move over the data to a separate directory, "version_resources",
and create links to the data in the repository. The first version of the data
is, by default, called "initial". Later on, we can add
further versions of the data, and switch between them.
}
\details{
@returns TRUE if the setup was successful
}
