% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_setup.R
\name{version_setup_first}
\alias{version_setup_first}
\title{Set up versioning for data in a resrepo for the first time}
\usage{
version_setup_first(quiet = FALSE, resources_path = NULL)
}
\arguments{
\item{quiet}{If TRUE, the user will not be prompted to backup their data.
Default is FALSE.}

\item{resources_path}{The path to the directory where the "version_resources",
the directory containing versioned data, will be stored. If NULL (the default),
"version_resources" is placed at the root of the repository.}
}
\value{
TRUE if the setup was successful
}
\description{
This function is used in a repository where data were being tracked and we
need to move them to a separate directory, "version_resources", and create
links to the data in the repository. The first version of the data is, by
default, called "initial".
}
\keyword{internal}
