% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_switch.R
\name{version_switch}
\alias{version_switch}
\title{Switch from one data version to another}
\usage{
version_switch(version)
}
\arguments{
\item{version}{The version to switch to}
}
\value{
TRUE if the switch was successful, FALSE otherwise
}
\description{
Change the soft links to \code{data/raw} and \code{data/intermediate} to point to the
specified version in \code{version_resources}
}
